Refactor entity serialization and state management by introducing SerializableFields for automatic dirty tracking. Update various extensions to utilize the new base class for improved field management. Remove deprecated game state properties and adjust buffer serialization methods to accommodate new encoding standards.

+++ packages/game-client/src/extensions/illuminated.ts
@@ -24,7 +24,7 @@ export class ClientIlluminated extends BaseClientExtension {
 
   public deserializeFromBuffer(reader: BufferReader): this {
     // Type is already read by the entity deserializer
-    this.radius = reader.readFloat64();
+    this.radius = reader.readUInt16();
     return this;
   }
 }
