+++ packages/game-server/src/extensions/ignitable.ts
@@ -53,6 +53,6 @@ export default class Ignitable implements Extension {
   }
 
   public serializeToBuffer(writer: BufferWriter): void {
-    writer.writeUInt32(encodeExtensionType(Ignitable.type));
+    writer.writeUInt8(encodeExtensionType(Ignitable.type));
   }
 }
