+++ packages/game-server/src/extensions/updatable.ts
@@ -45,6 +45,6 @@ export default class Updatable implements Extension {
   }
 
   public serializeToBuffer(writer: BufferWriter): void {
-    writer.writeUInt32(encodeExtensionType(Updatable.type));
+    writer.writeUInt8(encodeExtensionType(Updatable.type));
   }
 }
